+++ src/js/profileStats.js
@@ -1,0 +1,13 @@
+/*
+const STATS = document.querySelectorAll('.profile-stats-sidebar .profile-stat');
+var stats = [];
+
+for (let i = 0, len = STATS.length; i < len; i++) {
+    stats[i] = STATS[i].cloneNode(true);
+    stats[i].querySelector('div').remove();
+    stats[i] = stats[i].innerHTML.replace(/(\r\n|\n|\r|\$)/gm,"");
+
+}
+
+console.log(stats);
+*/