+++ src/js/raffle.js
@@ -23,6 +23,8 @@ var raffleMsg = document.querySelectorAll('.raffle-message');
 for (let i = 0, len = raffleMsg.length; i < len; i++) {
     raffleMsg[i].innerHTML = BBCODE_EXTRAS[0](raffleMsg[i].innerHTML);
     raffleMsg[i].innerHTML = BBCODE_EXTRAS[1](raffleMsg[i].innerHTML);
+    // Kappa
+    BBCODE_EXTRAS[2](raffleMsg[i]);
 }
 
 /* Link comments by repply (mouse hover) */
@@ -98,6 +100,8 @@ if (comBox) {
             // Urls
             str = BBCODE_EXTRAS[0](str);
             str = BBCODE_EXTRAS[1](str);
+            // Kappa
+            BBCODE_EXTRAS[2](newCom);
             // colors
             str = BBCODE[2](str);
             // text formatting
@@ -119,6 +123,8 @@ if (comBox) {
         // Urls ( not whitelisteds )
         allComs[i].innerHTML = BBCODE_EXTRAS[0](allComs[i].innerHTML);
         allComs[i].innerHTML = BBCODE_EXTRAS[1](allComs[i].innerHTML);
+
+        BBCODE_EXTRAS[2](allComs[i]);
     }
 
     comBox.addEventListener('DOMSubtreeModified', DebugBBCODEScrapTF);